Add MainForm view and methods
