+++ Clipboarder/Clipboarder.csproj
@@ -112,6 +112,7 @@
     <Compile Include="ClipboardMonitor.cs">
       <SubType>Component</SubType>
     </Compile>
+    <Compile Include="Content.cs" />
     <Compile Include="HotKey.cs" />
     <Compile Include="IMainDataLayer.cs" />
     <Compile Include="MainForm.cs">
@@ -119,6 +120,13 @@
     </Compile>
     <Compile Include="MainForm.Designer.cs">
       <DependentUpon>MainForm.cs</DependentUpon>
+    </Compile>
+    <Compile Include="MainFormPresenter.cs" />
+    <Compile Include="OldMainForm.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="OldMainForm.Designer.cs">
+      <DependentUpon>OldMainForm.cs</DependentUpon>
     </Compile>
     <Compile Include="PasswordBox.cs">
       <SubType>Component</SubType>
@@ -141,6 +149,9 @@
     </EmbeddedResource>
     <EmbeddedResource Include="MainForm.resx">
       <DependentUpon>MainForm.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="OldMainForm.resx">
+      <DependentUpon>OldMainForm.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Properties\Resources.resx">
       <Generator>ResXFileCodeGenerator</Generator>
